requirements.txt
affine==2.3.0
asgiref==3.3.1
attrs==20.3.0
branca==0.4.2
bumpversion==0.5.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
checksumdir==1.2.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
coloredlogs==15.0
colorlog==4.7.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
Django==2.2.9
django-rest-swagger==2.2.0
djangorestframework==3.12.2
docutils==0.16
dynaconf==3.0.0
Fiona==1.8.18
Flask==1.1.2
folium==0.12.1
GeoAlchemy2==0.6.3
geojson==2.5.0
geopandas==0.9.0
gitchangelog==3.0.4
gpxpy==1.4.2
grass==0.1.2
humanfriendly==9.1
idna==2.10
imageio==2.9.0
invoke==1.5.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
more-itertools==8.7.0
munch==2.5.0
MyApplication==0.1.0
networkx==2.5
numpy==1.19.5
openapi-codec==1.3.2
opencv-python==4.5.1.48
ortools==7.8.7959
pandas==1.1.5
Pillow==8.1.0
protobuf==3.15.3
psycopg2-binary~=2.8.6
pycparser==2.20
pyparsing==2.4.7
pyproj==3.0.0.post1
python-box==5.3.0
python-dateutil==2.8.1
python-log-indenter==0.9
python-telegram-bot==12.8
pytz==2021.1
pyvisgraph==0.2.1
PyWavelets==1.1.1
PyYAML==5.4.1
rapclust==0.1.2
rasterio==1.2.0
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
snuggs==1.4.7
sortedcontainers==2.3.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
statistics==1.0.3.5
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
tornado==6.1
tqdm==4.59.0
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1


